Alphabet challenge rewrite.
The repository should now contain separate folders to signify
the separation of the different systems.

The idea is to create a game for each letter in the alphabet
but to keep things simple I will have all games, platform code
and the engine code in the same repository instead of multiple ones.

See this as the initial commit of trying to convert my asteroids
game from linux in c to x-platform in c++.